requirements.txt
screen-brightness-control
appopener
quote
PyAutoGUI
openai
groq
selenium
webdriver-manager
googlesearch-python
cohere
requests
bs4
pygame
edge-tts
PyQt5
appopener
python-dotenv
pillow
pywhatkit
rich
mtranslate
keyboard
pyttsx3
nara